Final modem refactoring.
