+++ MMDVMHost.h
@@ -1,5 +1,5 @@
 /*
- *   Copyright (C) 2015-2020 by Jonathan Naylor G4KLX
+ *   Copyright (C) 2015-2021 by Jonathan Naylor G4KLX
  *
  *   This program is free software; you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -58,7 +58,7 @@ public:
 
 private:
   CConf           m_conf;
-  IModem*         m_modem;
+  CModem*         m_modem;
   CDStarControl*  m_dstar;
   CDMRControl*    m_dmr;
   CYSFControl*    m_ysf;
