+++ ModemPort.cpp
@@ -16,22 +16,8 @@
 *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 */
 
-#ifndef MMDVMModemPort_H
-#define MMDVMModemPort_H
+#include "ModemPort.h"
 
-class IMMDVMModemPort {
-public:
-	virtual ~IMMDVMModemPort() = 0;
-
-	virtual bool open() = 0;
-
-	virtual int read(unsigned char* buffer, unsigned int length) = 0;
-
-	virtual int write(const unsigned char* buffer, unsigned int length) = 0;
-
-	virtual void close() = 0;
-
-private:
-};
-
-#endif
+IModemPort::~IModemPort()
+{
+}
