+++ ModemSerialPort.cpp
@@ -1,5 +1,5 @@
 /*
-*   Copyright (C) 2016,2020 by Jonathan Naylor G4KLX
+*   Copyright (C) 2016,2020,2021 by Jonathan Naylor G4KLX
 *
 *   This program is free software; you can redistribute it and/or modify
 *   it under the terms of the GNU General Public License as published by
@@ -21,7 +21,7 @@
 #include <cstdio>
 #include <cassert>
 
-IModemSerialPort::IModemSerialPort(IModem* modem) :
+IModemSerialPort::IModemSerialPort(CModem* modem) :
 m_modem(modem)
 {
 	assert(modem != NULL);
